backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
firebase-admin==6.4.0
pillow==10.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
aiofiles==23.2.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cloudinary==1.36.0
requests==2.31.0
email-validator==2.3.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
ultralytics==8.1.11
